+++ CHANGELOG.md
@@ -64,6 +64,7 @@ With #144 you can now join streams with differing group by dimensions.
 - [#270](https://github.com/influxdata/kapacitor/issues/270): The HTTP server will now gracefully stop.
 - [#300](https://github.com/influxdata/kapacitor/issues/300): Add OPTIONS method to /recording endpoint for deletes.
 - [#304](https://github.com/influxdata/kapacitor/issues/304): Fix panic if recording query but do not have an InfluxDB instance configured
+- [#289](https://github.com/influxdata/kapacitor/issues/289): Add better error handling to batch node.
 
 ## v0.10.1 [2016-02-08]